Reorganise Cucumber Suite
